+++ src/client/app/data.ts
@@ -1,4 +1,4 @@
-import { InjectionKey, Ref, ref, readonly, computed, inject } from 'vue'
+import { InjectionKey, Ref, shallowRef, readonly, computed, inject } from 'vue'
 import { Route } from './router'
 import serializedSiteData from '@siteData'
 import { resolveSiteDataByRoute, PageData, SiteData } from '../shared'
@@ -20,7 +20,7 @@ export interface VitePressData<T = any> {
 // site data is a singleton
 export type SiteDataRef<T = any> = Ref<SiteData<T>>
 
-export const siteDataRef: Ref<SiteData> = ref(parse(serializedSiteData))
+export const siteDataRef: Ref<SiteData> = shallowRef(parse(serializedSiteData))
 
 function parse(data: string): SiteData {
   return readonly(JSON.parse(data)) as SiteData
@@ -46,9 +46,11 @@ export function initData(route: Route): VitePressData {
     frontmatter: computed(() => route.data.frontmatter),
     lang: computed(() => site.value.lang),
     localePath: computed(() => {
-      const { locales, lang } = site.value
-      const path = Object.keys(locales).find((lp) => locales[lp].lang === lang)
-      return withBase((locales && path) || '/')
+      const { langs, lang } = site.value
+      const path = Object.keys(langs).find(
+        (langPath) => langs[langPath].lang === lang
+      )
+      return withBase(path || '/')
     }),
     title: computed(() => {
       return route.data.title