+++ src/node/config.ts
@@ -2,17 +2,31 @@ import path from 'path'
 import fs from 'fs-extra'
 import chalk from 'chalk'
 import globby from 'globby'
-import { normalizePath, AliasOptions, UserConfig as ViteConfig } from 'vite'
+import {
+  normalizePath,
+  AliasOptions,
+  UserConfig as ViteConfig,
+  mergeConfig as mergeViteConfig
+} from 'vite'
 import { Options as VuePluginOptions } from '@vitejs/plugin-vue'
-import { SiteData, HeadConfig, LocaleConfig } from './shared'
+import {
+  SiteData,
+  HeadConfig,
+  LocaleConfig,
+  createLangDictionary
+} from './shared'
 import { resolveAliases, APP_PATH, DEFAULT_THEME_PATH } from './alias'
 import { MarkdownOptions } from './markdown/markdown'
+import _debug from 'debug'
 
 export { resolveSiteDataByRoute } from './shared'
 
-const debug = require('debug')('vitepress:config')
+const debug = _debug('vitepress:config')
+
+export type { MarkdownOptions }
 
 export interface UserConfig<ThemeConfig = any> {
+  extends?: RawConfigExports
   lang?: string
   base?: string
   title?: string
@@ -22,27 +36,28 @@ export interface UserConfig<ThemeConfig = any> {
   locales?: Record<string, LocaleConfig>
   markdown?: MarkdownOptions
   /**
-   * Opitons to pass on to @vitejs/plugin-vue
+   * Opitons to pass on to `@vitejs/plugin-vue`
    */
   vue?: VuePluginOptions
   /**
    * Vite config
    */
   vite?: ViteConfig
-  customData?: any
 
   srcDir?: string
   srcExclude?: string[]
 
   /**
-   * @deprecated use `srcExclude` instead
+   * Enable MPA / zero-JS mode
+   * @experimental
    */
-  exclude?: string[]
-  /**
-   * @deprecated use `vue` instead
-   */
-  vueOptions?: VuePluginOptions
+  mpa?: boolean
 }
+
+export type RawConfigExports =
+  | UserConfig
+  | Promise<UserConfig>
+  | (() => UserConfig | Promise<UserConfig>)
 
 export interface SiteConfig<ThemeConfig = any> {
   root: string
@@ -57,6 +72,7 @@ export interface SiteConfig<ThemeConfig = any> {
   markdown: MarkdownOptions | undefined
   vue: VuePluginOptions | undefined
   vite: ViteConfig | undefined
+  mpa: boolean
 }
 
 const resolve = (root: string, file: string) =>
@@ -66,22 +82,7 @@ export async function resolveConfig(
   root: string = process.cwd()
 ): Promise<SiteConfig> {
   const userConfig = await resolveUserConfig(root)
-
-  if (userConfig.vueOptions) {
-    console.warn(
-      chalk.yellow(`[vitepress] "vueOptions" option has been renamed to "vue".`)
-    )
-  }
-  if (userConfig.exclude) {
-    console.warn(
-      chalk.yellow(
-        `[vitepress] "exclude" option has been renamed to "ssrExclude".`
-      )
-    )
-  }
-
   const site = await resolveSiteData(root, userConfig)
-
   const srcDir = path.resolve(root, userConfig.srcDir || '.')
 
   // resolve theme path
@@ -90,22 +91,33 @@ export async function resolveConfig(
     ? userThemeDir
     : DEFAULT_THEME_PATH
 
+  // Important: globby/fast-glob doesn't guarantee order of the returned files.
+  // We must sort the pages so the input list to rollup is stable across
+  // builds - otherwise different input order could result in different exports
+  // order in shared chunks which in turns invalidates the hash of every chunk!
+  // JavaScript built-in sort() is mandated to be stable as of ES2019 and
+  // supported in Node 12+, which is required by Vite.
+  const pages = (
+    await globby(['**.md'], {
+      cwd: srcDir,
+      ignore: ['**/node_modules', ...(userConfig.srcExclude || [])]
+    })
+  ).sort()
+
   const config: SiteConfig = {
     root,
     srcDir,
     site,
     themeDir,
-    pages: await globby(['**.md'], {
-      cwd: srcDir,
-      ignore: ['**/node_modules', ...(userConfig.srcExclude || [])]
-    }),
+    pages,
     configPath: resolve(root, 'config.js'),
     outDir: resolve(root, 'dist'),
     tempDir: path.resolve(APP_PATH, 'temp'),
     markdown: userConfig.markdown,
     alias: resolveAliases(themeDir),
     vue: userConfig.vue,
-    vite: userConfig.vite
+    vite: userConfig.vite,
+    mpa: !!userConfig.mpa
   }
 
   return config
@@ -117,16 +129,53 @@ export async function resolveUserConfig(root: string): Promise<UserConfig> {
   const hasUserConfig = await fs.pathExists(configPath)
   // always delete cache first before loading config
   delete require.cache[require.resolve(configPath)]
-  const userConfig: UserConfig | (() => UserConfig) = hasUserConfig
-    ? require(configPath)
-    : {}
+  const userConfig: RawConfigExports = hasUserConfig ? require(configPath) : {}
   if (hasUserConfig) {
     debug(`loaded config at ${chalk.yellow(configPath)}`)
   } else {
     debug(`no config file found.`)
   }
+  return resolveConfigExtends(userConfig)
+}
 
-  return typeof userConfig === 'function' ? userConfig() : userConfig
+async function resolveConfigExtends(
+  config: RawConfigExports
+): Promise<UserConfig> {
+  const resolved = await (typeof config === 'function' ? config() : config)
+  if (resolved.extends) {
+    const base = await resolveConfigExtends(resolved.extends)
+    return mergeConfig(base, resolved)
+  }
+  return resolved
+}
+
+function mergeConfig(a: UserConfig, b: UserConfig, isRoot = true) {
+  const merged: Record<string, any> = { ...a }
+  for (const key in b) {
+    const value = b[key as keyof UserConfig]
+    if (value == null) {
+      continue
+    }
+    const existing = merged[key]
+    if (Array.isArray(existing) && Array.isArray(value)) {
+      merged[key] = [...existing, ...value]
+      continue
+    }
+    if (isObject(existing) && isObject(value)) {
+      if (isRoot && key === 'vite') {
+        merged[key] = mergeViteConfig(existing, value)
+      } else {
+        merged[key] = mergeConfig(existing, value, false)
+      }
+      continue
+    }
+    merged[key] = value
+  }
+  return merged
+}
+
+function isObject(value: unknown): value is Record<string, any> {
+  return Object.prototype.toString.call(value) === '[object Object]'
 }
 
 export async function resolveSiteData(
@@ -142,6 +191,6 @@ export async function resolveSiteData(
     head: userConfig.head || [],
     themeConfig: userConfig.themeConfig || {},
     locales: userConfig.locales || {},
-    customData: userConfig.customData || {}
+    langs: createLangDictionary(userConfig)
   }
 }